feat(solana): adds Message
