+++ packages/solana/programs/debridge-reporter/Cargo.toml
@@ -20,3 +20,4 @@ idl-build = ["anchor-lang/idl-build"]
 anchor-lang = "0.30.1"
 debridge-solana-sdk = "1.0.1"
 slots = { path = "../../shared/slots" }
+message = { path = "../../shared/message" }
