+++ packages/solana/shared/message/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "message"
+version = "0.1.0"
+description = "Message struct to use within Hashi"
+edition = "2021"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+
+[dependencies]
+anchor-lang = "0.30.1"
